Move search box at search page to the separate component
